api_structure/requirements.txt
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
azure-monitor-opentelemetry>=1.0.0
opentelemetry-instrumentation-fastapi>=0.41b0
apscheduler>=3.10.0
pytz>=2023.3
